+++ test/test_build.py
@@ -4,7 +4,7 @@
 import torch
 import torchtext.data
 
-from common.torchtext_test_case import TorchtextTestCase
+from .common.torchtext_test_case import TorchtextTestCase
 
 
 class TestNestedField(TorchtextTestCase):
@@ -26,3 +26,41 @@ class TestNestedField(TorchtextTestCase):
             sources, vectors='glove.6B.50d',
             unk_init=torch.nn.init.normal_, vectors_cache=".vector_cache")
 
+
+class TestDataset(TorchtextTestCase):
+    def test_csv_file_no_header_one_col_multiple_fields(self):
+        self.write_test_ppid_dataset(data_format="csv")
+
+        question_field = torchtext.data.Field(sequential=True)
+        spacy_tok_question_field = torchtext.data.Field(sequential=True, tokenize="spacy")
+        label_field = torchtext.data.Field(sequential=False)
+        # Field name/value as nested tuples
+        fields = [("ids", None),
+                  (("q1", "q1_spacy"), (question_field, spacy_tok_question_field)),
+                  (("q2", "q2_spacy"), (question_field, spacy_tok_question_field)),
+                  ("label", label_field)]
+        dataset = torchtext.data.TabularDataset(
+            path=self.test_ppid_dataset_path, format="csv", fields=fields)
+        expected_examples = [
+            (["When", "do", "you", "use", "シ", "instead", "of", "し?"],
+             ["When", "do", "you", "use", "シ", "instead", "of", "し", "?"],
+             ["When", "do", "you", "use", "\"&\"",
+              "instead", "of", "\"and\"?"],
+             ["When", "do", "you", "use", "\"", "&", "\"",
+              "instead", "of", "\"", "and", "\"", "?"], "0"),
+            (["Where", "was", "Lincoln", "born?"],
+             ["Where", "was", "Lincoln", "born", "?"],
+             ["Which", "location", "was", "Abraham", "Lincoln", "born?"],
+             ["Which", "location", "was", "Abraham", "Lincoln", "born", "?"],
+             "1"),
+            (["What", "is", "2+2"], ["What", "is", "2", "+", "2"],
+             ["2+2=?"], ["2", "+", "2=", "?"], "1")]
+        for i, example in enumerate(dataset):
+            self.assertEqual(example.q1, expected_examples[i][0])
+            self.assertEqual(example.q1_spacy, expected_examples[i][1])
+            self.assertEqual(example.q2, expected_examples[i][2])
+            self.assertEqual(example.q2_spacy, expected_examples[i][3])
+            self.assertEqual(example.label, expected_examples[i][4])
+
+        # 6 Fields including None for ids
+        assert len(dataset.fields) == 6
